Added docker-compose, Dockerfile, and fixed pylint issues with app.py.
