+++ README.md
@@ -1,3 +1,6 @@
 # Containerized App Exercise
 
 Build a containerized app that uses machine learning. See [instructions](./instructions.md) for details.
+
+# Running
+`docker-compose up --build`
