+++ compose.yaml
@@ -1,0 +1,7 @@
+version: '3'
+
+services:
+  web:
+    build: .
+    ports:
+      - "8000:5000"
