+++ requirements.txt
@@ -1,0 +1,6 @@
+flask
+python-dotenv
+pymongo
+hume
+pytest
+pylint
