--- src/machine-learning-client/readme.txt
@@ -1,1 +1,0 @@
-Code related to the machine learning client goes in this folder.
