--- src/web_app/readme.txt
@@ -1,1 +1,0 @@
-Code related to the web app goes in this folder.
